make function more generic and move it to testutils
